Migrated query_builder functions from leetcode_util.py to new query_builder.py
